Move ActiveSupport instrumentation bridge to Rails layer.
Keep core instrumentation standalone; register the AS::Notifications
adapter only from sidekiq/rails.rb when Rails is present.
